+++ src/ralph/backends/lrs/mongo.py
@@ -13,6 +13,7 @@ from ralph.backends.data.mongo import (
     MongoQuery,
 )
 from ralph.backends.lrs.base import (
+    RELATED_AGENTS_FIELDS,
     AgentParameters,
     BaseLRSBackend,
     BaseLRSBackendSettings,
@@ -80,7 +81,14 @@ class MongoLRSBackend(BaseLRSBackend[MongoLRSBackendSettings], MongoDataBackend)
         if params.statement_id:
             mongo_query_filters.update({"_source.id": params.statement_id})
 
-        MongoLRSBackend._add_agent_filters(mongo_query_filters, params.agent, "actor")
+        if params.related_agents:
+            MongoLRSBackend._add_related_agent_filters(
+                mongo_query_filters, params.agent
+            )
+        else:
+            MongoLRSBackend._add_agent_filters(
+                mongo_query_filters, params.agent, "actor"
+            )
         MongoLRSBackend._add_agent_filters(
             mongo_query_filters, params.authority, "authority"
         )
@@ -121,10 +129,49 @@ class MongoLRSBackend(BaseLRSBackend[MongoLRSBackendSettings], MongoDataBackend)
         )
 
     @staticmethod
+    def _get_agent_filters(
+        agent_params: AgentParameters,
+        target_field: Union[str, tuple[str, ...]],
+    ) -> Union[dict, None]:
+        if not agent_params:
+            return None
+        if not isinstance(agent_params, dict):
+            agent_params = agent_params.model_dump()
+
+        if not isinstance(target_field, str):
+            target_field = ".".join(target_field)
+
+        if agent_params.get("mbox"):
+            key = f"_source.{target_field}.mbox"
+            return {key: agent_params.get("mbox")}
+
+        if agent_params.get("mbox_sha1sum"):
+            key = f"_source.{target_field}.mbox_sha1sum"
+            return {key: agent_params.get("mbox_sha1sum")}
+
+        if agent_params.get("openid"):
+            key = f"_source.{target_field}.openid"
+            return {key: agent_params.get("openid")}
+
+        if agent_params.get("account__name"):
+            key_name = f"_source.{target_field}.account.name"
+            key_homepage = f"_source.{target_field}.account.homePage"
+            return {
+                "$and": [
+                    {key_name: agent_params.get("account__name")},
+                    {
+                        key_homepage: agent_params.get("account__home_page"),
+                    },
+                ]
+            }
+        return None
+
+    @classmethod
     def _add_agent_filters(
+        cls,
         mongo_query_filters: dict,
         agent_params: Union[AgentParameters, list[AgentParameters]],
-        target_field: str,
+        target_field: Union[str, tuple[str, ...]],
     ) -> None:
         """Add filters relative to agents to mongo_query_filters.
 
@@ -133,42 +180,39 @@ class MongoLRSBackend(BaseLRSBackend[MongoLRSBackendSettings], MongoDataBackend)
             agent_params (AgentParameters): Agent query parameters to search for.
             target_field (str): The target agent field name to perform the search.
         """
-
-        def _get_agent_filters(
-            _params: AgentParameters,
-        ) -> Union[dict, None]:
-            if not _params:
-                return None
-            if not isinstance(_params, dict):
-                _params = _params.model_dump()
-
-            if _params.get("mbox"):
-                key = f"_source.{target_field}.mbox"
-                return {key: _params.get("mbox")}
-
-            if _params.get("mbox_sha1sum"):
-                key = f"_source.{target_field}.mbox_sha1sum"
-                return {key: _params.get("mbox_sha1sum")}
-
-            if _params.get("openid"):
-                key = f"_source.{target_field}.openid"
-                return {key: _params.get("openid")}
-
-            if _params.get("account__name"):
-                key_name = f"_source.{target_field}.account.name"
-                key_homepage = f"_source.{target_field}.account.homePage"
-                return {
-                    key_name: _params.get("account__name"),
-                    key_homepage: _params.get("account__home_page"),
-                }
-            return None
-
         if not agent_params:
             return
         elif not isinstance(agent_params, list):
-            filters = _get_agent_filters(agent_params)
+            filters = cls._get_agent_filters(
+                agent_params=agent_params, target_field=target_field
+            )
             if filters:
                 mongo_query_filters.update(filters)
         else:
-            filters = [_get_agent_filters(params) for params in agent_params if params]
+            filters = [
+                cls._get_agent_filters(agent_params=params, target_field=target_field)
+                for params in agent_params
+                if params
+            ]
             mongo_query_filters.update({"$or": filters})
+
+    @classmethod
+    def _add_related_agent_filters(
+        cls, mongo_query_filters: dict, agent_params: AgentParameters
+    ) -> None:
+        """Add filters relative to agents to `where`, including any 'related agents'."""
+        if not agent_params:
+            return
+
+        related_filters = []
+        for field in RELATED_AGENTS_FIELDS:
+            field_filter = {}
+            cls._add_agent_filters(
+                mongo_query_filters=field_filter,
+                agent_params=agent_params,
+                target_field=field,
+            )
+            if len(field_filter) > 0:
+                related_filters.append(field_filter)
+        if len(related_filters) > 0:
+            mongo_query_filters.update({"$or": related_filters})