+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to
 - Backends: (internal) add support for multiple authority queries
 - API: allow users with new `authority/write` scope to override authority
        when PUTting or POSTing statements
+- API: implement GET '/statements' 'related_agents' query parameter
 
 ### Removed